refactor(media): extract shared audio decoder + SRT/VTT to src/lib
Move decodeToMono16k, srtTime, vttTime, buildSrt, buildVtt out of
AudioTranscriber into src/lib/media/decode-audio.ts and
src/lib/subtitles/cues.ts so the upcoming subtitle tool can reuse them.
AudioTranscriber now converts its Chunk[] to CueDoc at the call site;
its SRT/VTT output and existing tests are unchanged.
